Fix compilation errors

+++ src/fheroes2/gui/interface_itemsbar.h
@@ -113,7 +113,7 @@ namespace Interface
             ItemsIterator itemIter = items.begin();
             auto itemPosIter = itemPos.begin();
 
-            const bool useCustomCountX = ( _customItemsCountInRow.size() == tableSize.height );
+            const bool useCustomCountX = ( _customItemsCountInRow.size() == static_cast<size_t>( tableSize.height ) );
 
             for ( int32_t y = 0; y < tableSize.height; ++y ) {
                 const int32_t coutX = useCustomCountX ? _customItemsCountInRow[y] : tableSize.width;
@@ -285,7 +285,7 @@ namespace Interface
                 return;
             }
 
-            const bool useCustomCountX = ( _customItemsCountInRow.size() == tableSize.height );
+            const bool useCustomCountX = ( _customItemsCountInRow.size() == static_cast<size_t>( tableSize.height ) );
             const int32_t stepX = ( offsetBetweenItems.width + singleItemSize.width );
             const int32_t stepY = ( offsetBetweenItems.height + singleItemSize.height );
